requirements.txt
# Data manipulation
pandas
xlrd

# Authentication and environment variables
requests
python-dotenv

# Google Cloud Platform
google-cloud-storage
gcsfs
google-auth

# API
streamlit
